+++ c/py_datatable.h
@@ -101,7 +101,14 @@ DECLARE_GETTER(
 DECLARE_METHOD(
   window,
   "window(row0, row1, col0, col1)\n\n"
-  "Retrieve datatable's data within a window")
+  "Retrieve DataTable's data within a window")
+
+DECLARE_METHOD(
+  to_scalar,
+  "to_scalar()\n\n"
+  "For a 1x1 DataTable return its data as a Python scalar value. This in \n"
+  "contrast with `window(...)` method which returns a list-of-lists \n"
+  "regardless of the shape of the underlying DataTable.")
 
 DECLARE_METHOD(
   check,